Add ExposedPort class with port/protocol fields, move exposed port config-related parsing to frontend class (#449)
