+++ jib-core/src/integration-test/java/com/google/cloud/tools/jib/builder/BuildStepsIntegrationTest.java
@@ -21,6 +21,7 @@ import com.google.cloud.tools.jib.cache.CacheDirectoryCreationException;
 import com.google.cloud.tools.jib.cache.CacheDirectoryNotOwnedException;
 import com.google.cloud.tools.jib.cache.CacheMetadataCorruptedException;
 import com.google.cloud.tools.jib.cache.Caches;
+import com.google.cloud.tools.jib.frontend.ExposedPortsParser;
 import com.google.cloud.tools.jib.image.ImageReference;
 import com.google.cloud.tools.jib.registry.LocalRegistry;
 import java.io.IOException;
@@ -56,7 +57,9 @@ public class BuildStepsIntegrationTest {
             .setTargetImage(ImageReference.of("localhost:5000", "testimage", "testtag"))
             .setMainClass("HelloWorld")
             .setJavaArguments(Collections.singletonList("An argument."))
-            .setExposedPorts(Arrays.asList("1000", "2000-2002/tcp", "3000/udp"))
+            .setExposedPorts(
+                ExposedPortsParser.parse(
+                    Arrays.asList("1000", "2000-2002/tcp", "3000/udp"), logger))
             .setAllowHttp(true)
             .build();
 
@@ -100,7 +103,9 @@ public class BuildStepsIntegrationTest {
             .setTargetImage(ImageReference.of(null, "testdocker", null))
             .setMainClass("HelloWorld")
             .setJavaArguments(Collections.singletonList("An argument."))
-            .setExposedPorts(Arrays.asList("1000", "2000-2002/tcp", "3000/udp"))
+            .setExposedPorts(
+                ExposedPortsParser.parse(
+                    Arrays.asList("1000", "2000-2002/tcp", "3000/udp"), logger))
             .build();
 
     Path cacheDirectory = temporaryCacheDirectory.newFolder().toPath();
